move check variants
