+++ benchmarks/src/main/scala/zio/stm/SemaphoreBenchmark.scala
@@ -30,7 +30,7 @@ class SemaphoreBenchmark {
   @Benchmark
   def tsemaphoreContention() =
     unsafeRun(for {
-      sem   <- TSemaphore(fibers / 2L).commit
+      sem   <- TSemaphore.make(fibers / 2L).commit
       fiber <- ZIO.forkAll(List.fill(fibers)(repeat(ops)(sem.withPermit(STM.succeed(1)).commit)))
       _     <- fiber.join
     } yield ())
